solve calc_cosine

+++ week2/wpo1/ackermann.h
@@ -1,6 +1,6 @@
 // https://dodona.be/en/courses/5451/series/63959/activities/942621401
 
-long ackermann(short m, short n) {
+long ackermann(long m, long n) {
     if (m == 0) {
         return n + 1;
     } else if (m > 0 && n == 0) {
